+++ vpr/SRC/route/rr_graph_area.c
@@ -548,7 +548,7 @@ static float trans_per_buf(float Rbuf, float R_minW_nmos, float R_minW_pmos) {
 
 		num_stage = nint(log10(R_minW_nmos / Rbuf) / log10(4.));
 		num_stage = max(num_stage, 1);
-		stage_ratio = pow(R_minW_nmos / Rbuf, 1. / (float) num_stage);
+		stage_ratio = pow((float)(R_minW_nmos / Rbuf), float( 1. / (float) num_stage));
 
 		Rstage = R_minW_nmos;
 		trans_count = 0.;
@@ -587,7 +587,7 @@ static float trans_per_mux(int num_inputs, float trans_sram_bit,
 		/* This is a large multiplexer so design it using a two-level multiplexer   *
 		 * + 0.00001 is to make sure exact square roots two don't get rounded down  *
 		 * to one lower level.                                                      */
-		num_second_stage_trans = (int)floor(sqrt(num_inputs) + 0.00001);
+		num_second_stage_trans = (int)floor((float)sqrt((float)num_inputs) + 0.00001);
 		pass_trans = (num_inputs + num_second_stage_trans) * pass_trans_area;
 		sram_trans = (ceil(
 				(float) num_inputs / num_second_stage_trans - 0.00001)
